+++ src/app/Components/Skills/skills.style.css
@@ -62,6 +62,7 @@ div {
 .SkillContainer__Skills__item {
   height: 80%;
   width: 15%;
+
   margin-top: 20px;
   display: flex;
   flex-direction: column;
